+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,11 +1,5 @@
 package ru.yandex.practicum.filmorate.service;
 
-import java.util.ArrayList;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -18,16 +12,20 @@ import ru.yandex.practicum.filmorate.storage.interfaces.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.interfaces.UserStorage;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @Component
 @Service
 @Slf4j
 public class UserService {
-
     private final UserStorage storage;
-
     private final FilmStorage filmStorage;
 
     @Autowired
@@ -57,12 +55,19 @@ public class UserService {
         return result;
     }
 
-    public void delete(int userId) {
+    public User delete(Integer userId) {
+        log.info("Request to delete the user by ID = " + userId + " received");
+        if (userId == null) {
+            throw new NotFoundException("User with ID = " + userId + " not found");
+        }
+        if (userId < 0) {
+            throw new NotFoundException("User with ID = " + userId + " not found");
+        }
         if (getById(userId) == null) {
             throw new NotFoundException("User with ID = " + userId + " not found");
         }
-        log.info("Deleted film with id: {}", userId);
-        storage.delete(userId);
+        log.info("Deleted user with id: {}", userId);
+        return storage.delete(userId);
     }
 
     public User getById(Integer id) {
@@ -73,7 +78,6 @@ public class UserService {
     public void addFriend(Integer userId, Integer friendId) {
         checkUser(userId, friendId);
         storage.addFriend(userId, friendId);
-
         log.info("Friend successfully added");
     }
 
@@ -129,7 +133,6 @@ public class UserService {
         }
         log.info("For user {} recommended {} films from user {}.", userId, recommendations.size(),
                 idUserWithSameLikes.get());
-
         return recommendations;
     }
